store ui
modularised code
